Change block_locator_indexes height parame to size_t, style,

+++ include/bitcoin/bitcoin/block.hpp
@@ -49,11 +49,8 @@ BC_API bool operator==(
 
 BC_API uint64_t block_value(size_t height);
 BC_API hash_number block_work(uint32_t bits);
-
 BC_API hash_digest hash_block_header(const block_header_type& header);
-
-BC_API index_list block_locator_indexes(int top_height);
-
+BC_API index_list block_locator_indexes(size_t top_height);
 BC_API block_type genesis_block();
 
 } // namespace libbitcoin
